Allow to switch accounts under disabled group (#857)
* feat: Bump mobile version

* feat: Skip screenshots for ios

We do not change screenshots so often.

* feat: Enable interaction with disabled group account switch

Click were passing through account switches when their group was
disabled,because the account switch was disabled. We decided that
it was better for the switch to be still switchable, but in a grey
color to show the fact that its group is disabled, and that it is
not shown on the curve.