feat(bar): Handle bar theme overrides
For customization needs, we need to be able to pass overrides to
`cozy.bar.setTheme`. Do be able to override it easily, I added a map in
its own file. This way, we can override this single file in every
customization. In the original Banks we don't want any overrides, so it
is an empty object.

See https://github.com/cozy/cozy-bar/pull/473 for the implementation of
the overrides in the cozy-bar side.