feat: Debounce setting state for settings with 3s delay for perf
Setting state for settings woud re-render all the BalancePage, and this would delay the animation of the panel. Debounce has been introduced, settings save
happens 3s after the last toggling of a pnel.

Debounce is done at Balance, and not on each toggle, otherwise, debounce
would be only per panel.