chore(travis): Don't cache yarn
The yarn cache seems to be very big. Disabling it makes the cache goes
from ~1GB to ~200MB. The Travis doc warns about big cache archives:

"If you store archives larger than a few hundred megabytes in the cache,
it is unlikely that you’ll see a big speed improvement." (see
https://docs.travis-ci.com/user/caching/#how-does-caching-work)

I think the time to download/unpack this cache was longer than
installing dependencies without it, since the build time went from ~12mn
with it to ~8mn without it.