chore: Base services build on common webpack config
The services target was built using only the `webpack.target.services`
config file, creating a specific branch in our configs. This was not
very suitable with the customizations. Actually, in customizations the
service was built with no override at all. The goal is to make it easier
to use the overrides also to build the services.