refactor: Plugin strategy for client dependent objects
Events decouple the instantiation of the client
from the handlers that have to be plugged to events
from the client.

In other words, this means that we can extract the code
necessary to be run on {login,logout,refreshToken} for
objects dependent on cozyClient (here, the bar and the
push plugin) and put it in other files.

The plugin API (registerPlugin) and the plugin being
only a function that can attach events to the cozyClient
is inspired by webpack.