feat(notifications): Update TransactionGreater filter criteria
We consider that a transaction is new if its revision is lower than or
equals to 2. This change is necessary because the transactions are
categorized after their first save in the database. So when the
notifications service kicks in, no transaction will ever be in a state
in which its `rev` equals to 1