fix: Ensure cozy bar is initialized before rendering the app
By using the cozy-bar as a cozy-client plugin, we rely on the
cozy-client event system. But with it, we have no control over the order
in which the callbacks of each event are executed. This is a problem
because since the app's home page (/balances) needs to manipulate the
cozy-bar (changing its theme), we need to ensure that the cozy-bar is
initialized before rendering this page. To achieve that, we stop using
the cozy-bar as a cozy-client plugin. We use cozy-authentication
`MobileRouter`'s `onAuthenticated` prop and we initialize the cozy-bar
in it.