fix(Linker): Use realisationDate in byDate filter
We use `realisationDate` in priority for every transactions since this
is the most accurate date for each account types. This is particulary
true for credit card accounts, for which transactions always have a date
at the last day of the month. For other types, the `realisationDate` is
very close to the `date`, so we can use it safely in the matching
context.