feat(billsMatching): MatchFromTransactions handles matchingCriterias
We can't directly handle matching criterias since they are carried by
the bills and we only have the transactions. But the worst case is a
repayment schedule that generally covers 1 year. By looking for bills
one year before and one year after the oldest and the most recent
transactions, we should cover the vast majority of cases.