refactor: Separate templating of notification in two phases
- Notification class gains more responsibility -> children classes are
simpler
  - Does the two-phase templating
  - Gets data from children
  - Gets title/pushContent from children

Children classes of Notification are not involved in templating (apart
that they can define helpers).
Notification is the only class entrypoint for templating.

Some of the data preparation before rendering the template was done
inside each notifications/html/<notification-name>-html.js. It was moved
inside each Notification class.

For test purposes, the `fetchData` method was introduced in each Notification
class and is mocked inside tests with JSON data.