fix: Use the good format when calling services
We used a format that is working but that is not the one that is really
expected by the stack. It resulted in a weird job structure in the
io.cozy.jobs couch base and Bender was not able to show the service
name.

See https://github.com/cozy/cozy-stack/issues/2159