fix(loan): Don't crash on reload
When reloading the page, the accounts were not present, so when we tried
to access to the group's accounts relationship, the app crashed. To fix
this, I connected the component to the accounts query and ensured we
have the data before showing the accounts list. While data is loading,
a spinner is shown.