fix: Handle conflicts when saving settings
If the user tried to modify the settings after the budget notification had
been modified by the budget notification job, we could end up with a
document update error.

In this PR, a generic conflict resolution merge is implemented. The
schema of the doctype prioritize for some attributes the server version
and for other attributes the user version. This policies should end up
at the end in cozy-client.