fix: Limit warning is different for each account
With the introduction of rules, we had forgotten to update the
warningLimit code. Since the warningLimit is no longer global, we have
to lower the warning props down into the AccountRow component. To
facilitate checking if an account is below one of its limits, we build
a warningLimitPerAccount object (memoized with createSelector) that takes
care of merging all the account/group/unqualified rules down to a
warning limit per account