fix(Balance): Virtual accounts switches state
The getPanelsState function was looking for a property called `id`. On
regular accounts it exists, but not on virtual accounts. Since in the
same function we use `_id` for groups, I updated it to also use `_id` on
accounts. But I also added an `id` property on virtual accounts, so they
have the same shape as a real account from the stack.