feat: Add median absolute deviation measure
The median absolute deviation is similar to the standard deviation since
it is a measure of a series' dispersion but it is more robust against
outliers.

This statistical measure has been introduced to address the "gap"
problem in a bundle series. Sometimes there are gaps inside a recurrent
expense (anniversary month, missing data). The standard deviation was
very sensitive to these gaps. The MAD should be more robust to those
artefacts.

https://en.wikipedia.org/wiki/Median_absolute_deviation