feat: Ability to put a recurrence in finished status
A recurrence can be either "ongoing" or "finished". If it is finished,
we will not try to add transactions to it.

On desktop, the status is written in a dropdown button and actions on
the bundle like renaming are separated in another menu, whereas on
mobile, all the actions on the bundle are in the same menu.