chore: Update cozy-bar
In beta since https://github.com/cozy/cozy-bar/pull/688 needs to
be merged

I had to remove the use of the ProvidePlugin since it loads the bar
before our utils/react-exposer has the time to put React on the window.
This meant that the bar and the app did not share the same React which
causes problem in newer version of cozy-ui/cozy-client that use hooks.