fix: Send notifications emails with a content
Since the introduction of cozy-scripts in the banks repository, we use
the TerserPlugin from cozy-scripts. Previously, we would configure this
plugin not to mangle function names in the webpack.target.services file.
Now, we need to directly tweak the options of cozy-scripts to achieve
the same goal.

If function names are mangled, mjml registers incorrectly its components
and the rendered email is blank.