fix: Refresh the DisconnectedAccountModal when an account is deleted
Previously, the bank accounts objects, in case they were disconnected
from their io.cozy.accounts (when the io.cozy.accounts has been
deleted by the user), would be stored in the state. Storing them in
the state prevented them to be updated in the DisconnectedAccountModal
when one was deleted.

Here we store only the connectionId in the state, and we get the accounts
by filtering them (on their connectionId) from the accounts coming
from the store. This ensures we do not show a stale UI.