fix: Handle io.cozy.bank.accounts without connection relationship
Historically, the bank connectors would not link the
io.cozy.bank.accounts they retrieved with the io.cozy.accounts that ran
the connector. We did not support this usecase. This is done by grouping
those unconnected accounts by konnector slug. This is not optimal as
ideally, we would like those to be grouped by io.cozy.accounts and not
by io.cozy.konnectors. But since we cannot know the io.cozy.accounts,
it is the best we can do.