feat: Use the same AnalysisTabs component for analysis pages
This is important when using MUI tabs otherwise we'll have an unmount
then a mount of the AnalysisTabs when switching page. This would prevent
the correct animation of the tab indicator.

Having a common page deal with this also has the nice property that the
AnalysisTabs are not implemented in every page; the TransactionHeader
does not have to check the params to know if it has to render the tabs.