feat: Use the TransactionsPage as part of the CategoriesPage
Context: When in the CategoriesPage, the user can select a category and
then a subcategory and then see the transactions.

Before, the switch between the TransactionsPage and CategoriesPage
was done at route level. Now that we need the CategoriesHeader shared
between all subpages of the categories, it's better that the Categories
page uses the TransactionPage.

For this to correctly work, we need to be able:

* to disable the header of the TransactionPage since it is already
  rendered
* to disable the padding top of the TransactionPage since it is already
  taken care of by the CategoriesHeader (this padding is necessary since
  the header is fixed)
* we need to hide the chart from the CategoriesHeader if we are on a
  subcategory

This is what is done in this commit.

Breadcrumbs handling code is removed from the TransactionHeader since it
does not have to deal with that anymore.