feat: Add estimated budget card
First version, uses recurrence to compute planned transactions with
respect to the current filteringDoc. Those transactions are summed and
added to the balance of the filtered accounts to get an estimated
future balance.

The useEstimatedBalance hook will be shared between the card on the
transaction page and the header card in the planned transactions page.