feat: Store accounts into operations
To be able to create planned transactions, we need to have an
account for those transactions. Since we create transactions from the
recurrence bundle, we need to store account ids that have "participated"
to this recurrence bundle.

The service that creates recurrences is updated and also the code that
creates a recurrence "manually" from a transaction.