refactor: Use JS to compute header height and set list padding top
This way, if we had elements to the header, like the legal mention, the
list is automatically padded with the right amount. This is a bit ugly
but it does the job. It was done at first because I could not find a
way to override a stylus file on an overrided app, thus I could not
override the padding-top in an app where the legal mention was shown.