fix: Correct page name is sent when closing transaction modal
Since we replace the last part of page name when opening modal, we must
take care to change it back when closing, we cannot use the parent page
mechanism since the last part was replaced and not appended. The last
part is not appended since we need to limit the number of page name
parts to 3 because of the way page names are processed later on.