fix: Correctly check if notifications are enabled
At some point notifications became arrays and we forgot to update this
code to handle them. Notifications were still working for people with
the healthBillLinked notification enabled, since this one is not in
the form of an array.

Now, the isNotificationEnabled supports both arrays and single object,
and this is tested.