fix: Scrolling to bottom of view when landing on transactions page
When changing month on the transaction page, we need to scroll past
the top load more button, so that the user only see its transactions.
This scroll code is also ran when arriving on the page; in this case
there is no top load more button, thus the load more button that was
found was the bottom load more button, and we would scroll to the
bottom, potentially loading another month.

To fix, we make a distinction between the top and bottom load more
button, so that we cannot query the wrong one.

Also, the calculations for scrolling past the loadMore button was a bit
off. It now takes into account the header height and bar height, this
should be more robust.