feat: Look inside the trigger for bank account update date
This fixes the fonctional bug where the account is shown not to have
been updated when its data has not changed. This is due to the fact that
the banking konnectors only update bank accounts when they have changed.
Thus the cozyMetadata.updatedAt attribute can be unchanged for a long
time for account without a lot of changes (savings account for example).

It could confuse users that expect to see the last synchronisation date,
not the date where the object has been changed, in the AccountRow.

Here, the trigger last success date is used to determine the last sync
date shown in the account row. We keep the logic to take the bank account
last update date, in case the user has uninstalled the konnector. If the
trigger exists, its last success date is prioritized over the account
update.