fix: Only need to store konnector triggers for konnectorAlerts service
Triggers are not only for konnectors and can be used for sharing / notes
etc... If we do not scope the query, we end up with all the trigger
states inside the io.cozy.bank.settings that tracks the konnector
trigger states, where we only need the states for konnector triggers. In
a given cozy, we stored 173 trigger states, and now we only store 23.