fix: Empty groups after adding a bank
The first [PR] (https://github.com/cozy/cozy-banks/pull/2010) does not correct the problem of empty groups but keeps the interface up to date. (a [number] (https://github.com/cozy/cozy-banks/issues/2009) is open in order to optimize real time.

Empty groups remained displayed because they were not monitored and therefore not updated when the automatic group creation job was completed.

So I added GROUP_DOCTYPE in the list of doctypes to watch