fix: Show correct icons in Accounts and settings page
The slug for show the konnector icon was provided by the createdByApp.

In the event that the nestor konnector is used for the first time with
a Boursorama account and after we connect with the Boursorama konnector
with the same account.

So the nestor icon was always displayed instead of icon from the
last connected konnector.

In this example it will be boursorama icon.

Now Accounts and settings/accounts page show correctly icons