fix: Correct redirect to account transactions page on load
If coming directly to the /balances/<accountId>/details URL, the user
was redirected to the balance page and filters were not set.
The URL would only work if the page was already loaded.

At first, a query is in "pending" state; checking for "loading" is not
sufficient to know if the query has been loaded once:
SetFilterAndRedirect did not wait for queries to have been loaded.

Instead of "manually" checking, we use hasQueryBeenLoaded. Thus,
SetFilterAndRedirect correctly waits until both accounts and groups
queries have been loaded.