feat: Add flag to hide close button
This is necessary due to a circular dep between drive and
banks.
With previous drive, back arrow at the top left of the viewer
does not work. So we will use the cross button of the dialog
to close it.
When drive has been published, we will set the flag to true
in production so that the cross is hidden