fix: Category change is reflected in the search page
Before, we would store complete transactions in the state of the search
page. After a recategorization, the transactions were not updated in
the state of the component. Here, we store only the transaction ids
in the results and we retrieve the transactions from the "live"
transactions col that is updated when an item inside is updated.