fix: Do not show transaction dialog on click inside action modal
Since React makes events bubble up even if there is a portal, we have
problems since I put an event handler on the transaction row (and not an
event handler on each of the individual children). Here, to check if the
click originates from a portal, I check if the target event is inside
(DOM speaking) the row.