feat: Use fined grain auto update options
Monthly transaction queries are not updated by other monthly transaction
queries that load more data. Having the update could trigger nasty
bugs: if you load March transactions then January transactions then
come back to March, there would be January transactions inside the
March query because of the query auto update done by cozy-client. This
would prevent the fetchMore logic to work correctly ,leading to February
transactions not to be loaded until we scroll past January transactions.

We keep "updates" and "remove". "Update" is important so that when we
change the category or any attribute of a transaction, the query should
refreshed (in the redux sense).