fix: Use queryAll to fetch all recurrence bundle transactions
Previously, we'd use the default 100 limit when fetching bundle
transactions. When more than 100 transactions were linked to bundles,
some recurrences were fetched with the wrong number of transactions,
possibly 0. If a bundle was fetched with 0 transactions, it would
be deleted.

To prevent this problem, we use queryAll to fetch all documents
regardless of the pagination limit. The limitBy is increased to 1000
since we are in the context of a service, this not a problem fetch
large document batches.