fix: HeaderLoadingProgress at wrong place on categories page mobile
HeaderLoadingProgress is now displayed by CategoriesHeader as it was
not possible to control correctly where to display the progress bar
if it was done in the categories page (since we need on mobile to
display the progress bar above the chart, whereas on desktop it needs
to be below the header).

The progress bar is also now hidden if the round spinner is displayed.
The progress bar is thus only shown if there is already some data
showing and we are loading new data.