fix: Send transaction greater and balance lower notifications
Recently, the code was changed to use DocumentCollection.fetchChanges
and the include_docs: true option was forgotten.

Without include_docs, docs are not fetched when retrieving changes, and
this leads to the return value of fetchChanges to have no documents.
Since notifications are sent based on changes, there were no more
notifications sent for transactionGreater/balanceLower.

See https://github.com/cozy/cozy-client/issues/966 to discuss about
the ergonomics of fetchChanges.