fix: Use fully loaded query infinite loop
fetchMore could be called multiple times in a row for several
reasons, and this would trigger banks freeze at some point. By making
sure that we do not call fetchMore() while it is already fetching()
and by reducing the number of dependencies on the effect, the bad
behavior (banks slowing down, many fetchMore requests) is not
reproduced.

Added tests also to solidify the behavior of useFullyLoadedQuery.