fix: Remove during tests warning
NestedSelect options are keyed by the title of the option
or by the key attribute of the option. Since both
options had the same title "Salaire", we had a warning
during tests because two react keys were the same.

To remove the warning, we can set "key" as the _id of each recurrence in
in the NestedSelect options.