fix: Wrong order of transactions for credit card accounts
Transactions should be re-sorted after coming from the stack since
we display the realisationDate for credit cards. If transactions are
sorted "only" by date, the order of displayed transactions seems wrong
since the date that the user sees is the realisationDate. For delayed
debit card accounts, the realisation date can be at the beginning of
the month whereas the date is at the end of the month.

After the switch to incremental queries, the wrong transactions were
used, here we switch to using transactions from before, transactions
that are sorted by displayed date.

The test has to be updated to account for that since now even in the UI
test, transactions are sorted by displayed date. Since only the top
transactions are really displayed, we could not click on transactions
that were at the bottom.