fix: Delayed credit card transactions weren't displayed in analysis page
Delayed credit card transactions happen at the very end of the month.
Here the selector, having only day and not the hour, would not pick
transactions of the last day (because 2021-01-31 < 2021-01-31T12:00,
the first date being the upper bound of the query, and the second date
the date of the transaction), and thus no transaction were visible in
the categories page.