fix: When changing account on "Categories", date selector takes an available value
When the date selector receives new options, it will call its onChange
prop if the value is not available in the new options. It calls the
onChange callback with the available option that is nearest to the
current value. This algorithm was already coded but was not called
when new options were received.