fix: When changing month, scroll to the top
Otherwise, if the user has scrolled in the transactions list, she
stays at the current scroll instead of being transported to the
latest transaction of the selected month.

Since when selecting a month, the latest transaction of the selected
month is at the top of the transaction list, we can go the top of
the list.