fix: Fetching transactions for a group works on mobile
The $or query would fail with Pouchdb, with a "no index is usable for
this query". Seems like the $or query is not well supported by Pouch.
Fortunately, the $in operator can replace the the $or here
and seems more performant. It seems that we should avoid the $or
operator for our queries to be compatible between Couch and Pouch.