fix: No need to clean group before removing it
we removed account from groups on group delete probably because of copy-paste from remove account logic https://github.com/cozy/cozy-banks/pull/1996. This causes a bug because updating the group before deleting it sends a request and changes the revision of the document. But on mobile we also do this deletion in Pouch, which will modify the document by adding _deleted to true, and thus also modifies the revision number. There are therefore two documents concerning the group, with different revision numbers, and during the Pouch<->Couch sync, it is the document that does not have the _deleted that is recovered. This makes the deleted group reappear.