fix: Use correct css var color for warning and success
introduced by https://github.com/cozy/cozy-banks/pull/2221
We didn't make all the changes to the BC in v49 of cozy-ui, but since the old version of the cozy-bar was using a version prior to v49 of cozy-ui this was not a problem. Now this is not the case anymore.