fix: No data displayed Analysis Page (iOS)
On iOS we had no data displayed in analysis page.in "all accounts" case.

It would seem that PouchDB does not manage the indexing of the `_id` field.
At each request executed with this field there are no data in return.

Finally, in this case we want just to get data from dates.