fix: Add ICU for Node 12 and earlier
That way we harmonize the test between node 13+
and earlier version.

Since Node 13, Node comes with ICU support, so
component forcing `toLocaleString` like Figure
component from UI behave differently between
Node 12 and Node 16.

Since we're in the process of migrating to
Node 16 but our production is still on Node 12
let's have the same behavior for 12 and 16 like
that we can run the same test.

By adding ICU we give node 12 the capacity to
understand thing like toLocalString...

So we just needed to update the snapshot since
now the Figure is displayed in French.

Be careful, for figure like: 1 234,00€ the
space is a special character.