Revert "fix(account): Update trigger without fetch policy"
This reverts commit 814965c01111beb4b63358048d1f4b168e9a3e13.

Performance problems have been detected, because too many requests
happens on the Balance page (Home page of Page).
Instead of requesting 3 times to the stack, to get the
`jobs/triggers?Worker=konnector&Type=@cron` resource, now this
will be requested once (thanks to the fetchPolicy).

The reverted commit used to refetch the konnectors triggers,
thanks to another commit inside cozy-harvest-lib, that was
reverted in cozy-harvest-lib:
https://github.com/cozy/cozy-libs/commit/
61d204c3caca04713dac7543b3e4913ae2513f17

So it was not useful anymore.

In order to fix the updated job status, when re-connecting a
connector inside banks, another commit will be added in this PR.