fix: Add RealtimeQueries on triggers
This makes the HarvestConnectorModal removes the error message
when a connector has been successfully reconnected.

Formerly this PR fixes the above problem:
https://github.com/cozy/cozy-banks/pull/2407

However the fix on cozy-harvest-lib has been removed by
https://github.com/cozy/cozy-libs/commit/61d204c3caca04713dac7543b3e4913ae2513f17

To keep a behaviour consistent with CozyHome, we enable the
realtime on triggers, to benefit of fresh data while triggers has
been updated.