feat: Add wait jobs
Wait jobs allow harvest to tell to banks it is waiting for a job to be
triggered by a BI webhook.

This wait job will displayed like a real job in progress but will be
replaced and removed by the real job with the same slug or after 5
minutes.

Harvest can trigger this wait jobs by sending a realtime notify on
io.cozy.jobs with the slug of the expected job.

Since these wait jobs are triggered before any account or trigger
creation, they don't have any associated account and will be replaced by
any job with the same expected slug