refactor: Balance doesn't need to know about triggers
Triggers are only used when there is no account, no need to retrieve
them every time. So we created a new component that handle the query
to triggers.

Also, instead of using the HOC withBanksSlug that makes the hooks to be
launched again, we store the data inside the BanksContext in order to be
able to use it without fetching anything.

this withBanksSlug should be removed, but this HOC is currently use in
class component, so it's hard to remove it ATM.