fix: Use RealTimeQuery from Cozy-Client
Mount the realtime query at the top of the App tree.
Like that, realtime is always handled and we don't have
to handle it manually when the balance component is mounted
or not.

Like that, we can add a fetchPolicy to the transactions that
never expires since realtime should handle it.