fix: Add and remove tags at the same time
We could not add one or more tags to a transaction,
and at the same time delete one or more tags already present.
The transaction was not up to date(_rev) for the second operation.

In addition to fetching the latest version of the transaction,
it needs to be hydrated so that it has the `tags`(hasManyTags) property