feat: Make the reimbursement page use the balance details header
This basically replaces the title in virtual account balance details
with the account switch dropdown.
This allows to navigate to other accounts, without having
to go back to the balance page first.
This also syncs the wording of the title in virtual account balance details
with the one used on the balance page.