feat: Update devtools notification content to try flagship redirection
Banks devtool is one of our simpliest way to try notifications.
I added three properties used by the flagship app to redirect to an app
instead of the home after opening the notification to try this new feature.

It changes nothing for banks because these three
parameters are not used by banks.