fix: BalanceGreater notifications
If the conditions defined in  “high balance” are met, then a notification is sent to the user to inform him that an account (or group of accounts) has reached this balance.
If the balance changes, and the condition is still met, then a 2nd notification is sent.

If the balance of the account (account group) changes, but the condition is maintained, then there are no new notifications.
If the balance of the account (account group) goes out of the condition, then the notification can be sent when the condition is met again.