feat: Add helper for make brands since query
In order to complete the list of connectors
present in the brands.json file.

We overwrite the maintenance information, name and regexp
with registry information as a priority.

The brands.json file currently remains mandatory,
because on the connector side we do not have information
such as `contact` or `health` and some `regexp`.

And add permission for call registry