fix: Pass down the client to make getBrands work
It fixes the issue with `window is undefined` but now I've having error
`can't call filter on undifined` because even if we `makeBrands` we don't
have the brands within the cozy-client store. So the getBrands() that
call store.getState().brands return undefined every time.

I don't think this is the right way to fix the issue. We should not pass
down the client like that. Also, we should not rely on window to get the
client. We need to think a better solution.