fix: Crash of onOperationOrBillCreate service due to `window`
This service no longer worked since the work on
`brands`, the cause coming from the call to the
`getBrands` function called in `getBillRegexp`.
We now pass the `brands` to him via his arguments.

This is a fix to get the service back up and running
while waiting for a better approach.